+++ pkg/cs3/client.go
@@ -1,4 +1,4 @@
-package svc
+package cs3
 
 import (
 	gateway "github.com/cs3org/go-cs3apis/cs3/gateway/v1beta1"
@@ -15,16 +15,12 @@ func newConn(endpoint string) (*grpc.ClientConn, error) {
 	return conn, nil
 }
 
-func getGatewayServiceClient(endpoint string) (gateway.GatewayAPIClient, error) {
+// GetGatewayServiceClient returns a new cs3 gateway client
+func GetGatewayServiceClient(endpoint string) (gateway.GatewayAPIClient, error) {
 	conn, err := newConn(endpoint)
 	if err != nil {
 		return nil, err
 	}
 
 	return gateway.NewGatewayAPIClient(conn), nil
-}
-
-// GetClient returns a gateway client to talk to reva
-func (g Graph) GetClient() (gateway.GatewayAPIClient, error) {
-	return getGatewayServiceClient(g.config.Reva.Address)
 }
